Completed ALL animations
I have now completed all of the animations for the coursework. These consist of the question mark animation and the star animations. If I have time I may add some animations to the congratulations screen so that the "WELL DONE" and "ANSWER" labels animate also, but this depends on how much time I have left.